requirements.txt
openai==0.25.0
Pillow==9.3.0
reportlab==3.6.12
requests==2.28.1
rich==12.6.0
python-dotenv==0.21.0